+++ packages/realm/index.node.js
@@ -1,6 +1,6 @@
 ////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2022 Realm Inc.
+// Copyright 2023 Realm Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import "./fs";
-import "./device-info";
+/* eslint-disable @typescript-eslint/no-var-requires -- We're exporting using CJS assignment */
+/* eslint-env commonjs */
 
-export * from "../../index";
+module.exports = require("./dist/bundle.node").Realm;